+++ examples/test-site/src/components/FlowContainer/index.tsx
@@ -19,6 +19,8 @@ import withContentfulTouts from './withContentfulTouts';
 import withRichTextVariations from './withRichTextVariations';
 import withSingleAccordionVariations from './withSingleAccordionVariations';
 import withImageVariations from './withImageVariations';
+import withIframeVariations from './withIframeVariations';
+import withYouTubeVariations from './withYouTubeVariations';
 
 import { asFlowContainerWithMargins } from './token';
 
@@ -30,6 +32,8 @@ const FlowContainerDefault = flow(
   withToutVariations,
   withContentfulTouts,
   withSingleAccordionVariations,
+  withIframeVariations,
+  withYouTubeVariations,
   asFlowContainerWithMargins,
   withMandatoryCategories(['Orientation', 'Type']),
 )(FlowContainer);