+++ packages/bodiless-richtext-ui/src/components/index.ts
@@ -16,3 +16,4 @@ export { default as Overlay } from './Overlay';
 export { default as CloseButton } from './CloseButton';
 export { default as ClickableWrapper } from './ClickableWrapper';
 export { default as PreviewWrapper } from './PreviewWrapper';
+export { default as TextSelectorWrapper } from './TextSelectorWrapper';