+++ packages/bodiless-richtext/src/RichTextContext.tsx
@@ -25,6 +25,7 @@ const defaultUI = {
   CloseButton: 'span',
   ClickableWrapper: 'button',
   PreviewWrapper: 'div',
+  TextSelectorWrapper: 'div',
 };
 
 export type UI = {
@@ -33,6 +34,7 @@ export type UI = {
   Overlay?: ComponentType<HTMLProps<HTMLDivElement>> | string;
   CloseButton?: ComponentType<HTMLProps<HTMLSpanElement>> | string;
   ClickableWrapper?: ComponentType<HTMLProps<HTMLButtonElement>> | string;
+  TextSelectorWrapper?: ComponentType<HTMLProps<HTMLDivElement>> | string;
 };
 
 export type PreviewUI = {